resolve variable types

+++ Analyzer/src/Analyze1Lep.cc
@@ -163,7 +163,7 @@ void Analyze1Lep::Loop(NTupleReader& tr, double, int maxevents, bool)
         {
             // Define Lumi weight
             const auto& Weight  = tr.getVar<float>("Weight");
-            const auto& lumi = tr.getVar<float>("Lumi");
+            const auto& lumi = tr.getVar<double>("Lumi");
             eventweight = lumi*Weight;
             
             const auto& eleLepWeight = tr.getVar<float>("totGoodElectronSF");
@@ -171,10 +171,10 @@ void Analyze1Lep::Loop(NTupleReader& tr, double, int maxevents, bool)
             const auto& muNonIso     = tr.getVar<float>("totNonIsoMuonSF");
             leptonweight = eleLepWeight*muLepWeight;
             
-            pileupWeight = tr.getVar<float>("puWeightCorr");
-            bTagWeight   = tr.getVar<float>("bTagSF_EventWeightSimple_Central");
+            pileupWeight = tr.getVar<double>("puWeightCorr");
+            bTagWeight   = tr.getVar<double>("bTagSF_EventWeightSimple_Central");
             htDerivedweight = tr.getVar<float>("htDerivedweight");
-            prefiringScaleFactor = tr.getVar<float>("prefiringScaleFactor");
+            prefiringScaleFactor = tr.getVar<double>("prefiringScaleFactor");
             
             weightQCDCR *= eventweight*muNonIso*prefiringScaleFactor*pileupWeight;
             weightNoHT *= eventweight*leptonweight*bTagWeight*prefiringScaleFactor*pileupWeight;
